requirements.txt
pandas 
snowflake-connector-python
pandas
more-itertools
sqlalchemy 
mysql.connector
pymysql